add trucate and copy table checkout_paid for payment & shipping actions

+++ my-app/app/lib/actions.tsx
@@ -10,7 +10,9 @@ import {
   sendMessage,
   shippingQuery,
   paymentQuery,
-  eraseQuery
+  eraseQuery,
+  //queryToPrepareTable, 
+  //queryToCopyTable
 } from './db';
 import { revalidatePath } from 'next/cache';
 import { AuthError } from 'next-auth';
@@ -289,6 +291,8 @@ export async function messageToSend(prevState: {message: string} | undefined, fo
   }
 }
 
+
+
 // shipping from order
 export async function shippingRequest(prevState: {message: string} | undefined, formData: FormData) {
   try {
@@ -306,10 +310,16 @@ export async function shippingRequest(prevState: {message: string} | undefined, 
         const request = await shippingQuery("INSERT INTO shipping VALUES (?, ?, ?, ?, ?, ?, ?)",
           [email, user, address, npa, phone, passwd, filterTotal]);
         if (request) {
-          const eraseTable = await eraseQuery("TRUNCATE TABLE cartorder");
-          if (eraseTable) {
-            revalidatePath("/order");
-            return {message: "Shipping done !"};
+          const prepareCopy = await queryToPrepareTable("TRUCATE TABLE checkout_paid");
+          if (prepareCopy) {
+            const copyTable = await queryToCopyTable("INSERT INTO checkout_paid * FROM cartorder");
+            if (copyTable) {
+              const eraseTable = await eraseQuery("TRUNCATE TABLE cartorder");
+              if (eraseTable) {
+                revalidatePath("/order");
+                return {message: "Shipping done !"};
+              }
+            }
           }
         }
       }
@@ -335,10 +345,16 @@ export async function paymentRequest(prevState: {message: string} | undefined, f
         const request = await paymentQuery("INSERT INTO payment VALUES (?, ?, ?, ?, ?)",
           [user, date, securitycode, checkcardValue, filterTotal]);
         if (request) {
-          const eraseTable = await eraseQuery("TRUNCATE TABLE cartorder");
-          if (eraseTable) {
-            revalidatePath("/order");
-            return {message: "Payment done !"};
+          const prepareCopy = await queryToPrepareTable("TRUCATE TABLE checkout_paid");
+          if (prepareCopy) {
+            const copyTable = await queryToCopyTable("INSERT INTO checkout_paid * FROM cartorder");
+            if (copyTable) {
+              const eraseTable = await eraseQuery("TRUNCATE TABLE cartorder");
+              if (eraseTable) {
+                revalidatePath("/order");
+                return {message: "Shipping done !"};
+              }
+            }
           }
         }
       }
